backend/requirements.txt
fastapi
uvicorn
motor
openai
httpx
reportlab
python-multipart
sentence-transformers
faiss-cpu
pymupdf
pytesseract
python-dotenv
psutil
tf-keras
